requirements.txt
aiosqlite==0.17.0
anyio==3.6.2
asyncpg==0.27.0
click==8.1.3
fastapi==0.89.1
h11==0.14.0
httptools==0.5.0
idna==3.4
iso8601==1.1.0
psycopg2-binary==2.9.5
pydantic==1.10.4
pypika-tortoise==0.1.6
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
sniffio==1.3.0
starlette==0.22.0
tortoise-orm==0.19.3
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
